Rewrite the test using Spek 2
Interestingly the previous JUnit test was not being run by the CI
though there ia other JUnit test (which do not seem to be run either).
The decision to write JUnit test was based on noticing a JUnit test
and thinking that is the convention for this project. It turns out
mostly there are Spek tests. Rewrite was also helpful with better
grouping, improving legibility and reuse within the test.
