+++ build.gradle
@@ -6,6 +6,8 @@ buildscript {
         coroutines_version = '0.22.1'
         android_gradle_version = '3.1.1'
         android_support_version = '27.1.1'
+        spek_version = '2.0.9'
+        legacy_spek_version = '1.1.5'
         mockito_version = '2.10.0'
         constraints_layout_version = '1.1.0'
         espresso_version = '3.0.1'
