+++ pusher-platform-core/src/test/kotlin/com/pusher/util/TokenProviderTestUtil.kt
@@ -1,0 +1,16 @@
+package com.pusher.util
+
+import com.nhaarman.mockito_kotlin.anyOrNull
+import com.nhaarman.mockito_kotlin.doReturn
+import com.nhaarman.mockito_kotlin.mock
+import com.pusher.platform.network.Futures
+import com.pusher.platform.tokenProvider.TokenProvider
+import elements.Error
+
+fun mockFailingTokenProvider(error: Error): TokenProvider = mock {
+    on { fetchToken(anyOrNull()) } doReturn failureAs(error)
+}
+
+private fun failureAs(error: Error) = Futures.schedule {
+    error.asFailure<String, Error>()
+}
